Reorganize classes and init files for submodule.
